feat：研发商店组件上架支持自定义版本号 #11545

+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/common/service/impl/StoreBaseCreateServiceImpl.kt
@@ -128,7 +128,7 @@ class StoreBaseCreateServiceImpl @Autowired constructor(
             creator = userId,
             modifier = userId,
             latestFlag = true,
-            busNum = CommonUtils.generateNumber(VersionUtils.getMajorVersion(version), 1, STORE_BUS_NUM_LEN).toInt()
+            busNum = CommonUtils.generateNumber(VersionUtils.getMajorVersion(version), 1, STORE_BUS_NUM_LEN)
         )
         val storeBaseExtDataPOs = StoreReleaseUtils.generateStoreBaseExtDataPO(
             extBaseInfo = storeBaseCreateRequest.extBaseInfo,
